func_code/requirements.txt
azure-functions
azure-keyvault-secrets
azure-identity
